Attempt of a *.webarchive extractor; improved distribution build script.

+++ build-dist.sh
@@ -42,28 +42,66 @@ local TARGET="${1}"
 function add_scripts()
 {
 local TARGET="${1}"
+local UPD="${2}"
+local NG="${3}"
+local CMD1
+local CMD2
 
-	TERM=xterm-color jython -c '
+	if [ -z "${NG}" ]; then
+		CMD1='jython'
+	else
+		CMD1='ng ng-jython'
+	fi
+
+	if [ -n "${UPD}" ]; then
+		CMD2='-f'
+	fi
+
+	TERM=xterm-color ${CMD1} -c '
 import import_site_packages
 import sqlite3
 import __run__
+#reload(import_site_packages)
+#reload(sqlite3)
+reload(__run__)
 '
 # Sadly, currently does not work
-#	cp '__run__$py.class' '__main__$py.class'
-	zip "${TARGET}" 'import_site_packages$py.class' '__run__$py.class' 'sqlite3$py.class'
+	cp '__run__$py.class' '__main__$py.class'
+	zip ${CMD2} "${TARGET}" 'import_site_packages$py.class' '__run__$py.class' 'sqlite3$py.class' '__main__$py.class'
 	rm -f 'import_site_packages$py.class'
 }
+
+
+
+while [ "${#}" -gt 0 ]; do 
+	case "${1}" in
+	-u)
+		shift
+		UPD=1
+	  ;;
+	-ng)
+		shift
+		NG=1
+	  ;;
+	*)
+	# Must be the TARGET filename
+		break
+	  ;;
+	esac
+done
 
 TARGET="${1}"
 
 if [ -z "${TARGET}" ]; then
-	echo "Usage: build-dist.sh target.jar"
+	echo "Usage: build-dist.sh [-u] [-ng] target.jar"
 	exit 1
 fi
 
-rm -f "${TARGET}"
-build_jar "/opt/local/share/java/jython" "${TARGET}"
-set_entrypoint "${TARGET}"
-add_scripts "${TARGET}"
+if [ -z "${UPD}" ]; then
+	rm -f "${TARGET}"
+	build_jar "/opt/local/share/java/jython" "${TARGET}"
+	set_entrypoint "${TARGET}"
+fi
+add_scripts "${TARGET}" "${UPD}" "${NG}"
 ls -lh "${TARGET}"
